requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.23.2
tensorflow==2.20.0
numpy>=1.26.0
pillow
python-multipart
requests
